Qt: Add simple Window class and PlatformWonderBrush
The window just consists of a UI thrown together with QtDesigner,
approximating the layout of the Haiku window.

Unfortunately with the release build the application crashes in X11 code
when trying to show the window. The debug build doesn't crash. Also some
code changes that should not have any effect make the release version
work.