Qt: Implement BHandler and BMessenger
We use Qt's event mechanism to deliver the message via a dedicated
PlatformMessageEvent to a proxy object of the BHandler which, if the
handler still exists, calls BHandler::MessageReceived(). The tricky part
is the object's thread affinity. The helper template classes
PlatformObjectHandler and PlatformWidgetHandler make sure the object is
assigned to the thread the derived class is assigned to, so that the
message is delivered in the correct thread.