Qt: TransformToolConfigView: Avoid notification loop
Parameter changes via the controls are synchronously propagated to the
tool. Changes in the tool state are asynchronously messaged to the
config
view. When processing the latter we suppress notifying the tool now, or
we'd otherwise risk a notification loop whenever the user manages to
perform another parameter change before the message for the previous one
arrived.