Paint also needs a regular copy constructor...
...since it is also used via SharedPaint and the paint cache. The default
copy constructor which seems to be compiler generated does not do the
right thing regarding the Gradient: It is not supposed to be shared by
multiple Paint instances. (Not currently, anyways.) Fixes a crash due to
a listener still being attached to the Gradient (the original Paint) when
the cloned Paint is destroyed.