Paths of a Shape object are visible in the ObjectTreeView
This required a number of changes which are likely not yet complete. And/or
partially broken. Shape no longer references Paths directly, but instead
contains PathInstances, which refer both the Path and the owning Shape. This
allows to click paths in the object tree view and learn the paths
transformation inherited from the shape. In theory, the PathStateTool could
edit paths which live in different Shapes, while it is required to handle
each path with its own transformation, depending on which shape it is
contained in.

The ObjectTreeView now references BaseObjects instead of Objects. The
drag&drop should currently be broken due to this. Paths can only be dragged
between Shapes, and mixed dragging of Paths and other (real) Objects should
not be possible, but is currently not prevented.